Refactor: Move utils to `model/utils` directory
- Updated import paths for helper functions from `utils` to `model/utils`
- Removed `./` export paths in `package.json` for `utils`
- Deleted `src/utils/index.ts` file as it's no longer needed
