+++ src/model/completion.ts
@@ -1,10 +1,10 @@
 import { type PartialDeep, type SetOptional } from 'type-fest';
 
-import { deepMerge, mergeEvents, type Prettify } from '../utils/helpers.js';
 import { createOpenAIClient } from './clients/openai.js';
 import { AbstractModel, type ModelArgs } from './model.js';
 import { type Model } from './types.js';
 import { calculateCost } from './utils/calculate-cost.js';
+import { deepMerge, mergeEvents, type Prettify } from './utils/helpers.js';
 
 export type CompletionModelArgs<CustomCtx extends Model.Ctx> = SetOptional<
   ModelArgs<
