+++ src/model/utils/helpers.ts
