--- src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './cache.js';
-export * from './helpers.js';
